Removed dynamically created types, support pickling
The Dynamic and Strong classes are now just used to enforce type
adherance, with the Dynamic "forcing" everything to be an `object`.

Note that upon pickling a yamlizable object, it will lose the round trip
data. It may be possible to somehow retain this, but it'll take some
digging into ruamel.yaml