Changed Attribute.has_default to only be true when unassigned
When not assigned, there will be no value in getattr(obj, storage_name).
getattr can just be called with self.default to allow the default to be
used when it has not been assigned. This way we can detect whether the
default value was explicitly assigned. Also, you can clear any value
that was explicitly set by using del